feat(lm-orchestrator/vector_search): prune vocab after update

+++ crates/lm-orchestrator/src/config.rs
@@ -62,7 +62,7 @@ pub struct Config {
     pub qdrant_collection_name: String,
 
     #[arg(env = "MAX_SPARSE_SIZE", long)]
-    pub max_sparse_size: u64,
+    pub max_sparse_size: u32,
 
     #[arg(env = "EMBEDDING_MODEL_PATH", long)]
     pub embedding_model_path: String,
